backend/requirement.txt
fastapi[all]
torch
torchvision
torchaudio
scikit-learn
numpy
pandas
uvicorn
python-dotenv